Implement SMPSO  algorithm to Evogil platform.
We adapt SMPSO algorithm base on the OMOPSO algorithm. The idea of the algorithm we have implemented is from the paper of A.J. Nebro et al where the SMPSO adopts a constriction coefﬁcient. The mechanism in such a way that the accumulated velocity of each variable j(in each particle) is further bounded by means of the following velocity constriction equation.

* Most modification on line 153  --> def compute_speed(self):